made some progress, but still not working correctly.
The issue is that lower level keystrokes are handled first by the method intercetptKeyBeforeQueueing,
so that wake up events can be handled. I have figured out that regardless of what you do to the parameters there, if you tell it not to handle it there, it passes the original keystroke to the other method,
interceptKeyBeforeDispatching. This seems to be handled in native methods out of my control.
The solution, I think, is to catch it in the first one, tell it not to handle it, then catch it again in the second and then change it. trying that next.