successfully changed bluetooth keyboard power button to forward delete by hijacking the point where it checks for fallback actions and changing it there.
there is still a lot of unneeded code that I will clean up.
I am also thinking there will be a problem if it is a key that apps normally handle (not power).
First, I stopped the system from handling the power key. Then it gets passed to the app, which cannot handle it.
then, I think it looks for a fallback, at which point I can change it and have it actually get passed to the app.

if the original key was one that it did not lookf ro a fallback for, I am not sure where I could change it.