removed some unneeded code and commented out logging.
confirmed my suspicions about it not always looking fro a fallback. If an ap handles the power key (or thinks it does, even if it does nothing)
then it never calls the method where I am able to switch it. For example, I cannot forward delete by my current method while remotely accessing my computer through splashtop,
though it does not lock the screen, so that is still a partial success.